Feat: Camera Fade In & Fade Out

+++ WinAPI_Hwanse/CPlayer.cpp
@@ -97,11 +97,12 @@ void CPlayer::render(HDC _dc)
 	bf.SourceConstantAlpha = 127;
 
 	AlphaBlend(_dc 
-		, vPos.x - width / 2.f
-		, vPos.y - height / 2.f
-		, width, height
+		, int(vPos.x - width / 2.f)
+		, int(vPos.y - height / 2.f)
+		, int(width), int(height)
 		, pTex->GetDC()
-		, 0, 0, width, height
+		, 0, 0
+		, int(width), int(height)
 		, bf);
 }
